+++ collatex-pythonport/HISTORY.rst
@@ -6,11 +6,12 @@ History
 2.0.0pre5 (2014-06-11)
 ++++++++++++++++++++++
 
-* Fixed handling of empty cells in JSON output.
+* Added JSON output to collate method.
 * Added option to collate method to enable or disable parallel segmentation.
+* Added table output to collate_pretokenized_json method, next to the already existing JSON output.
+* Fixed handling of empty cells in JSON output of pretokenized JSON.
 * Fixed compatibility issue when rendering HTML or SVG with IPython 2.1 instead of IPython 0.13.
 * Corrected RST syntax in package info description. 
-* Added JSON output to collate method.
 
 2.0.0pre4 (2014-06-11)
 ++++++++++++++++++++++